--- pgp/src/main/kotlin/idoubtthat/pgp/Keyservers.kt
@@ -1,11 +1,0 @@
-package idoubtthat.pgp
-
-import io.ktor.client.*
-import io.ktor.client.engine.cio.*
-
-class Keyservers {
-    val client = HttpClient(CIO) {
-        expectSuccess = true
-    }
-
-}
